Bug#37413349 Replace all use of 'utf8' and '_utf8' and 'collation_xxx utf8_xxx' in .mtr tests [noclose]
utf8 is currently an alias for utf8mb3. We want to change this some
time in the future.  Replace all use of 'utf8' and '_utf8' and
'collate utf8_xxx' and 'collation_connection = utf8_xxx' in .mtr
tests, and re-record .result files.

See patches for

Bug#35767731 	Replace all use of 'utf8' and '_utf8' and 'collate utf8_xxx' in .mtr tests

Change-Id: I739aade148a25214092d29b67995f58e3127adc6