Bug#37191149 Replace the DB path with DB Name in the handlerton drop_database_t API
This patch changes the signature of drop_database callback from the
'DB path' to 'DB name' because required schema name is already
available at call site.
Therefore, there is no need to pass the DB path and, then do the
conversion from path to name in the callback implementations.
Also fixes the cascade effect of the change in the API implementation
of NDB Storage Engine Implementation.

Change-Id: I08b250baac42b835ee3391c133b28f129596eeab