Bug#37417386 Add an interface library for xxhash.c
We currently compile xxhash.c into our own binaries in order to use
xxhash functions independently from the lz4 library (bundled or source).
This requires various INCLUDE_DIRECTORIES and COMPILE_DEFINITIONS
cmake directives in order to work.

Build an INTERFACE library for it instead, and link with that wherever
xxhash functions are used.

Change-Id: I4347ce4618860b3f367416bf72dda1675439d657