Bug#37271834: Hypergraph: Wrong results for query with semijoin with lateral dependencies [2/3, lateral, noclose]
The hypergraph optimizer may create invalid query plans for some
queries that have a semijoin inside a correlated derived table.

The problem is that the hypergraph optimizer doesn't track the lateral
dependencies of a correlated derived table if it's merged into the
outer query block, and it may therefore choose a join order that does
not make the lateral dependency tables available when the join in the
correlated derived table is performed.

Fixed by detecting and registering these lateral dependencies in
FindLateralDependencies().

Change-Id: I83b94dde835fad94fb130c004e7ecd9a59fffe7f