Bug#35861465: Hypergraph: Poor and inconsistent row estimates               for index merge
Row estimates for a query having an OR predicate in the WHERE clause
are not correct for hypergraph optimizer. This is because, while estimating
the number of output rows for a given table, it only considers estimates
from single index range scans and for the predicates that cannot use
such a scan, pre-determined selectivity is applied. However, we can do better.
Range optimizer would have set up index merges for OR predicates if
each part of the OR predicate can use a different range scan. This can be
used to get better estimates for these predicates.
To estimate rows from an index merge, we find the range scan with
the lowest estimate for each part of the OR predicate. Each of these
estimates are added to get the final estimate for the OR predicate.

Change-Id: I9d7a432d02b5d09230a8975ba40578de5917aaef