Bug#37199882: Hypergraph chooses ROR intersection for               a single range
If a secondary index has the same or a subset of the
keyparts of a clustered primary index and if the
query specified has equality ranges only on a subset
of the primary index keyparts, hypergraph optimizer
currently picks rowid ordered intersect plan using the
secondary index and the clustered primary index.
Since a index range scan or a index lookup on the primary
index would suffice in this case, using both these
indexes is an overkill.
There are a few problems that could be looked at here.
The first one being, the cost of the index merge scan
should have been higher than using the single index range
scan. This will be taken care as part of WL#16657.
As of now, cost for index merge scans are not exactly
comparable with index range scans as the cost model
for index range scan in hypergraph is changed. However
the cost of using index range scans as part of index
merge scans uses the old costing. This will be taken
care as part of the worklog.
The second problem is that the same keyparts are covered
by both secondary index and the primary index. So,
ideally when looking at the intersect plan, it should
have rejected because adding the clustered primary index
should not improve the selectivity and therefore the cost.
This works with indexes having keyparts which are
not part of primary index. The problem arises only for
clustered index keyparts, because while setting up
fields needed for the query which is used later to see
if a plan is covering, it does not add the fields
that are part of the clustered primary index. As a result,
for the failing query, when it first looks at the
secondary index, it concludes that it does not cover
any fields and therefore the plan is not covering. It then
adds the clustered primary index and checks the scan
selectivity. Since there is no information about the previous
scan covering any fields, it thinks that adding primary
index improves selectivity and therefore the cost. So we
multiply the selectivity for the same range twice resulting
in less number of output rows for the plan and therefore
the cost.
This issue occurs only when the condition has fields from
primary index. If it uses any field that is not part of
the primary index, fields needed for the query will have
such a field and therefore while finding the best order
to use the indexes for an intersect plan, it will order
the secondary index which has that field first, followed
by other indexes. And while setting up the plan for this
intersect order, it concludes the plan is covering as
soon as it sees the secondary index and exits.
We do not see this with old optimizer because when adding a
clustered primary index at the end, it only compares the
cost of using the secondary indexes in the plan with the
best cost estimate so far for the table. Although the
final cost for the index merge does include the cost of
using the clustered primary index, its not used for
comparison. Not sure as to why we do this. But this
actually makes the intersect plan costlier compared
to the range scan on primary key with the old optimizer.
Therefore it picks the range scan.

Solution for now, is to not add a secondary index range
scan to an intersect plan if it has subset of keyparts
from a clustered primary index when the clustered
primary index can itself be used.
For the case where we have only one range, avoiding
creation of plans with index merge should suffice. There
is no point in creating index merge plans for single ranges.
A more proper fix would be to look into the handling of
indexes having primary index keyparts while choosing
the intersect order, while getting the scan selectivity,
while picking the final plan etc. We are not doing this
for now.

Change-Id: I294faf0e2032f35e60bf70c2457c39b26a29ae00