Bug#37466984 MAX() as a Window Function Blindly Returns NULL for the First Row under Specific Conditions
Optimized computation for MAX(col) with ORDER BY DESC col didn't work as expected.

We have a table with two int rows (1 and 2 respectively), and the window function

  MAX(c0) OVER (ORDER BY c0 DESC ROWS BETWEEN 1 FOLLOWING AND 1 FOLLOWING)

row one:  c0: 2 frame has one row, i.e. row two, expect wf result: 1
row two:  c0: 1 frame is empty,                  expect wf result: NULL

Hence, the row based frame for row one (value 2 due to DESC) has one
row i.e. the value 1, and the frame for row two is empty.  This works
correctly for MIN, but not for MAX.

Btw, the work-around described (adding of 0) in the report, disables
the optimization used for MIN/MAX when the argument is the same as the
expression ordered by, hence the optimized path is not used.

In Item_sum_hybrid::compute():

Setting of variable:

   const bool visiting_first_in_frame =
        (m_window->optimizable_row_aggregates() &&
         m_window->rowno_being_visited() ==
             m_window->first_rowno_in_rows_frame()) ||
        !m_window->optimizable_row_aggregates();

didn't get the correct value (we would expect true for the first row
in the row based frame), in the repro row 2. The reason is that this
predicate failed:

  m_window->rowno_being_visited() == m_window->first_rowno_in_rows_frame()

because m_first_rowno_in_rows_frame has not been set, but had the
default value 1, whereas m_window->rowno_being_visited() had the value
2. In this case, the correct value for m_first_rowno_in_rows_frame is
also 2 for the first row in the frame, hence the error.

Fix: make sure m_first_rowno_in_rows_frame gets set in the appropriate
place.

Change-Id: I2d158d9bdd0cde44b6e594c98d4537f53b3b3f6a