Bug#37453301 close_connection_after_refresh=0 opens new connections
With WL#16652 the connection of the metadata-cache should stay
established, but checking the audit-log of the server shows that:

- the metadata-cache connection stays open (correct)
- ... and a new connections from the metadata-cache are created (wrong)

Background
==========

The metadata-cache uses one connection which can go to any of the
cluster nodes as they are SELECTs only.

Some statements require send an UPDATE and must therefore go to the
PRIMARY. For those statements new connections where opened.

Change
======

- if the cached connection of the metadata-cache goes to the PRIMARY,
  use it for the UPDATE queries too.
- otherwise, open a new connection for each of the updates.

Change-Id: I4ee470cb30b11f4288d1f5cce09e3c4185677ed5