Bug#35570336 Cluster/J should allow the JVM to handle SIGPIPE
Deprecate the function ndb_mgm_set_ignore_sigpipe(), and remove
a call to ignore SIGPIPE from TransporterFacade::start_instance().

Add two new functions in ndb_socket.h: ndb_socket_can_disable_sigpipe()
and ndb_socket_disable_sigpipe(). "Disabling" is possible on Windows
(where the OS will never raise SIGPIPE on a TCP socket anyway) and on
BSD platforms that support the socket option SO_NOSIGPIPE. It is not
possible on Linux platforms.

On platforms where "disabling" is possible, use it for SocketClient,
SocketServer, and Transporter sockets.

On Linux, in ndb_mgm_connect(), add a check using sigaction() to see
whether the default signal disposition has been overridden. Only ignore
the signal if the default handling is still in place. This automatic
check is intended to take the place of ndb_mgm_set_ignore_sigpipe();
that is the reasoning behind the deprecation.

Change-Id: I7f976018698f3416fd66c7259c1d64789c2d580d