Bug#37287708 - Inappropriate HTTP method for logout
Description
===========
As an authenticated MRS user, I expect to be able to log out and close the session o a given
REST service by sending an HTTP request to the /${service_name}/authentication/logout endpoint.
Although one might argue this is an idempotent operation, it isn't, from the HTTP standpoint,
"Safe" (https://developer.mozilla.org/en-US/docs/Glossary/Safe/HTTP i.e. it potentially alters
the state of the server (if there was, in fact, a session to begin with).
So, it shouldn't the "GET" method (https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods/GET
which requires the operation to be Safe.

Fix
===
The endpoint should handle requests using the "POST" method instead of the "GET" method.

Change-Id: I9ab97e5ab0b65ad21429f5c2f535ac29b6c6674c