Bug#37461433 Changes to 'mrs_user' table are not detected by mrs plugin
When the user has authenticated in MRS the metadata entry for that user from
mrs_user table is cached and the cached data is used for further authentications.
This is true even if the user has been disabled or removed from the metadata.

This patch changes the behavior. Now the audit-log is monitored for changes in
mrs_user table. In case od DELETE or UPDATE the user is invalidated in MRS
cache to force the update from metadata.

Change-Id: Iabcbf1e03b73370dd2d402d8b8cb195ce4e99dc7