Bug#37413349 Replace all use of 'utf8' and '_utf8' and 'collation_xxx utf8_xxx' in .mtr tests [mysqldump_57,noclose]
utf8 is currently an alias for utf8mb3. We want to change this some
time in the future.  Replace all use of 'utf8' and '_utf8' and
'collate utf8_xxx' and 'collation_connection = utf8_xxx' in .mtr
tests, and re-record .result files.

See patches for

Bug#35767731    Replace all use of 'utf8' and '_utf8' and 'collate utf8_xxx' in .mtr tests

Change-Id: I266e01dec01a7c00a8abb0a16b22659c84fa19b1