Bug#37413349 Replace all use of 'utf8' and '_utf8' and 'collation_xxx utf8_xxx' in .mtr tests [Cutf8,noclose]
utf8 is currently an alias for utf8mb3. We want to change this some
time in the future.  Replace all use of 'utf8' and '_utf8' and
'collate utf8_xxx' and 'collation_connection = utf8_xxx' in .mtr
tests, and re-record .result files.

This patch fixes usage command-line options character-set-server and
collation-server.

See patches for

Bug#35767731    Replace all use of 'utf8' and '_utf8' and 'collate utf8_xxx' in .mtr tests

Change-Id: I9c75365da068b894555a883d05ee1eb3deeff789