Bug#36421710 mysql server 8.3.0 crashed at Query_block::nest_derived [9.3 version]
We hit an empty pointer in Query_block::nest_derived used by
Query_block::transform_scalar_subqueries_to_join_with_derived.

[ 8.0 version: CUBE not supported.
  8.4 version: We implicitly try to
  do a prepare for secondary engine in the original repro due to GROUP
  BY CUBE(x), since CUBE is not supported by the primary engine. In the
  simplified repro, we explicitly enable the transform to see the
  error.
  9.3 version: repro gives error message due to outer reference in
  window's ORDER BY clause. If that is removed, the error doesn't occur. ]

At one stage we try to transform a query with a subquery shown as the
following by the optimizer trace:

/* select#2 */
select `t1`.`i`
from <constant table> join
     `cte2`
     on (/* select#5 */ select 3 from `cte2`) <> 0

i.e. we have a join condition, but only one table (cte2) in the from
list. This is the result of an earlier transformation.  However,
Query_block::nest_derived is not prepared for this scenario,
presupposing that there be at least two tables when a join condition is
involved: accessing Table_ref::embedding only works if we have an
embedding, i.e.  an inner table, but here we have only a single table.

Solution: test for this situation, which allows the transform to
succeed.

Note: in 8.4 the subquery can't be offloaded to secondary engine
anyway, since there are other subqueries that are not eliminated, but
will work for primary engine for the simplified repro (no CUBE) with
transform enabled.

Change-Id: If83e7a8e72372a1d3da30b9155ede594352a2845