Bug#35940224: Hypergraph: Semijoin with rand() gives wrong result
Some semijoins with non-deterministic semijoin conditions gave wrong
results in the hypergraph optimizer.

The problem was that the semijoin condition was evaluated too high up
in the join tree. This had the effect that the query returned only the
rows that matched the first row in the IN subquery, instead of
returning all the rows that matched any of the rows in the IN
subquery. The query result therefore contained fewer rows than it
should have.

The root issue is that the semijoin transformation places the semijoin
condition in the WHERE clause. The optimizer tries to push the
semijoin conditions down to the proper place in the join tree, but the
hypergraph optimizer failed to do this for certain non-deterministic
semijoin conditions.

Fixed by making the hypergraph optimizer try a bit harder to push down
semijoin conditions from the WHERE clause.

Change-Id: Idfa2f03d9ff7af89c5361a76a2924e905ac67fb9