Bug#36606073: json_table and exists make a bug
When JSON_TABLE was used in an IN or EXISTS subquery, and referenced a
table in the outer query block, wrong results were returned in some
cases when semijoin transformation had been performed.

The problem was that the semijoin transformation didn't call
fix_after_pullout() on the argument of the JSON_TABLE expression, so
that it still appeared to have an outer reference. Because of this,
the table function had no lateral dependency on the outer table, and
the optimizer could end up choosing a plan with the outer table inner
to the table function.

Fixed by implementing fix_after_pullout() on Table_function, and
calling it from fix_tables_after_pullout().

Change-Id: I9e3de5710486c0af24fb9d5177a2f69cd195d330