Bug#37500484 MRS does not check for the metadata schema upgrade while reading from the metadata.
The MRS should always check for the current version while doing the metadata
refresh. If it discovers that `0.0.0` version is set it should treat it as a
metadata upgrade in progress.

This patch changes the MRS behavior. Now when it discovers the upgrade it stops
the service and restores it when the version is back compatible with the MRS.

Change-Id: Ie3b8219e9447dfd55ab19fef164f90898f0c769a