Bug#37507319: Unnecessary allocation of optimizer trace buffer in hypergraph optimizer
The hypergraph optimizer always allocates an UnstructuredTrace object
for storing optimizer trace, even when tracing is disabled. This hurts
multi-threaded point select performance, and should be avoided when
it's not needed.

Even though the UnstructuredTrace object is allocated on the stack and
is left empty, a contained std::deque object performs heap
allocations, so it has a non-negligible cost.

Fixed by allocating an UnstructuredTrace object only if optimizer
tracing is enabled.

Change-Id: I201977839a02a79aecec8c27e81ef8a678ff436e