Bug#37509068: Hypergraph allocates strings for optimizer trace labels when not tracing [1/2, QUALIFY, noclose]
The hypergraph optimizer allocates strings for some optimizer trace
labels, even if tracing is not enabled. This is unnecessary, and it
hurts performance, especially for simple queries such as point
selects.

This patch skips allocating trace labels for the post-window filter if
tracing is not enabled or the query has no post-window filter.

Change-Id: I5ab7d10ac1484aa38fc04eb19235696695612800