Bug#36593244 MySQL server crashes at AggregateIterator::Init()
Analysis: Server exit due to missing error check: we have an ORDER BY
containing an aggregation on a set operation that was not caught
because it was hiding inside a subquery, where the aggregate was
aggregating on the outer query block, i.e. the set operation, not in
its own query block. The subquery was optimized away, being replaced
by an integer constant, so the check for has_aggregation failed on the
original check site, in setup_order.

Solution: Move check for whether the current Item_sum is an aggregate
in a block belonging to a set operation to check_sum_func. This also
enabled a previously failing query in union.test; the new check is more
precise.

Change-Id: I49df7f327b0544d658dafa2f2223155683f6de73