Bug#37168132: Unexpected creation of an large index, causing an               inaccessible table
Problem

The ALTER TABLE operation using the INPLACE algorithm to modify
a column's size (when an index is defined on it) may lead to the
creation of an index exceeding the valid size limit (i.e. greater
than 767 bytes) for tables with a row format of Redundant or Compact.
This occurs if the row format is not explicitly specified during
table creation and the 'innodb_default_row_format' setting is
configured to Compact or Redundant.

Analysis

When a table is created without explicitly specifying the row format,
the value from the global system variable 'innodb_default_row_format'
is used. If this variable is set to Compact or Redundant, the maximum
allowed index length should be 767 bytes.

ALTER TABLE operation, calls
ha_innobase::prepare_inplace_alter_table_impl() to prepare and validate
the inplace ALTER TABLE operation. For existing indexes, the function
validates maximum supported key length(i.e 767 bytes) against the
index prefix length( 0 for non prefixed index) instead of the index
key part length. As a result, the function potentially allows the
creation of an index exceeding 767 bytes for tables with Compact or
Redundant row formats.

Fix

The function ha_innobase::prepare_inplace_alter_table_impl()
has been modified where the index key part length is validated
against the maximum supported index length using the function
innobase_check_index_len(). This ensures that ALTER TABLE,
INPLACE operation will report an error if the modification of
the column size(when an index is defined on it) will result in
an invalid index size.

Change-Id: I7d32f45da783bf8f6216bfb2070cccec40663f8a