BUG#37543598: Do not truncate long function names in demangled stack traces
* Background

The server comes with its own stack trace printer, which produces a
stack trace when the server crashes.

* Problem

In case a stack entry including the demangled function name exceeds
512 bytes, it gets truncated.  It does not even print the newline, so
multiple truncated stack entries appear concatenated on the same
output line. This makes the stack incomplete and hard to read.

* Analysis

The stack trace printer uses a home-grown simplified version of printf
to print each line. This function has an internal buffer of 512 bytes.

* Fix

Use the home-grown printf function only to print numbers and
delimiters which are short. For any any potentially long strings -
filenames and demangled functions - write them directly to the output.

Change-Id: Iec3f30c8feb929e9c3096fcc395d82d2ad7d415f