WL#16788,Bug#37560672 Built-in MySQL Task Instrumentation and Asynchronous Execution
Added support for mysqlTask option for procedure and function endpoints.
These endpoints can be configured to be executed asynchronously using the
mysql_tasks framework by setting the mysqlTask.driver option to "database"
or "router". If set to "database", the routine is executed at the server
in an event. If set to "router", it is executed at the router in a separate
thread. In both cases, the invoking PUT/POST call returns immediately.

GET is implemented for returning status of a task and DELETE to kill a task.

Also:

- made routine endpoints accept POST as an alias to PUT (bug#37560672)
- stats counters for kEntityCounterRestAffectedItems and
kEntityCounterRestReturnedItems were removed for routines, as their
implementation was incomplete and their expected meaning is unclear for
this case

Change-Id: Iab931b5a2c814a2debf4f269fa9604c29b744886