Bug#37521401: server not coming up for upgrade from 9.2.0 to 9.3.0
Issue 36890891 ("If SQL in constraint fails, table becomes
unopenable / unfixable") added functionality that on server
upgrade looks at all tables, and for each one that has
constraints, default expressions, partitioning, or virtual
columns, we try to open it (to see whether we still can,
so we can detect tables that can no longer be opened due
to changes in SQL syntax). Opening will however also fail
if the storage engine for that table is not available yet.
This patch therefore restricts the above feature to InnoDB
tables.

(backport to 8.0, 8.4)

Change-Id: I1d2b881b2b737ed2d83625cddf301230b4266dfb