Bug#37529202 mysqld has incomplete list of schema participants
Introduce a new status variable Ndb_schema_participant_count.
This is an integer indicating the number of mysql servers that
are participating in NDB schema change distribution.

Add a new MTR test schema_participant_count. Using the standard
my.cnf file in suite/ndb, this test expects to see two participants.
It will wait for up to one second for the count to reach two.

Change-Id: I9af947155f79d5869a172e4154a3d0163e0cba54