Bug#34685358 (1/3) Integrate crund build in our cmake files
This patch adds a new command-line option --start-and-test to
mysql-test-run.pl.  When the option is supplied, MTR will set the
variable MTR_SKIP_TEST_CLEANUP in the environment, run a single test
specified on the command line, and then leave all servers running
(as with --start) until interrupted by the user.

The SQL script supplied to create the CRUND test and result tables
is adapted to become a new MTR test ndb_big.crund.

This allows a developer to start a crund benchmark environment
by running the single command:

  mtr --start-and-test crund

The crund SQL scripts crund_create.sql and crund_drop.sql
in mysql-test/suite/ndb_big can also be used outside of MTR.

Change-Id: Iaf34b35c8624685d952de5256c21ecc4f06e30a5