Bug#37551514  drop_all_tabs tool fails to drop a table
Problem:
Tool is designed to drop all test tables defined in
NDBT_Tables.cpp. It iterates over all tables and drop all,
one by one. If a target table does not exist (either because
it has already been dropped or was never created) the tool
stops iteration and the remaining tables are not dropped.

Solution:
If dropTable fails with error '723: No such table existed'
for a specific table continue dropping the remaining tables.

Change-Id: I505f37037472c2b84afd03e8f70f34480cbe5d57