Bug #36993445 : virtual index unstable after rollback when index_id is                 greater than max uint32
SYMPTOM:
When performing an insert operation on a table with a virtual index,
re-inserting the same row after rolling back the previous insert causes
unexpected behaviour failure if the index_id is greater than or equal to 2^32.

ROOT CAUSE:
The function used to log and read the index_id for virtual indexes in
the undo record does not support 64-bit integers, causing issues when
the index_id exceeds the maximum value of a 32-bit unsigned integer.

FIX:
Replace the existing logging and reading functions with
mach_u64_write_much_compressed() and mach_read_next_much_compressed(),
which support 64-bit integers.

Change-Id: Iaa508b0f56b7103709a04424afc917371e46951f