Bug#37486661 ndb_metadata thread can deadlock with truncate table
Summary:

Prevent errors through lock corruption during TRUNCATE by not closing
and reopening the NDB shared table handler while the table is
recreated.

Problem:

When running continuous TRUNCATE SQL commands, while concurrently
performing a NDB Backup, the MySQL Server could hang or ultimately
crash. The hang and crash occurred around the same code point, at
thr_lock(..), respectively with `wait_for_lock` and `has_old_lock`.

Analysis:

The order of operations of the TRUNCATE command, at the NDB plugin
layer, can be narrowed to three main operations:

1. CLOSE table (ha_ndbcluster::close)
2. CREATE TABLE (TRUNCATE <=> DROP AND CREATE with ha_ndbcluster::create)
3. OPEN table (ha_ndbcluster::open)

By observing the patterns, it is seen that if the Table undergoing
TRUNCATE is still under BACKUP then the DROP fails and thus step (2)
returns early. This causes the NDB_SHARE, a similar table handle as
TABLE_SHARE, to continue to exist. On a success case, DROP would cause
NDB_SHARE to be marked as dropped (ref count = 0) and CREATE would
create a new NDB_SHARE. On a failed DROP case, the original NDB_SHARE
is not marked as dropped.

At this point, the OPEN step (3) is run and one of the operations is
to initialize the thr_lock data for the handler, linking it with the
NDB_SHARE thr_lock, and setting it into TL_UNLOCK. But, the NDB_SHARE
and the handler's thr_lock data was LOCK'ed (specifically
TL_WRITE_ALLOW_WRITE) therefore the aforementioned operation (set to
TL_UNLOCK) effectively tramples the current lock data.

The result is a badly cleaned and "unlocked" lock. The subsequent
operations on that Table that require SQL locks might find the two
problematic scenarios:

1. Found old lock with an owner and will wait thus, but that owner has
already exited and trampled the lock

2. Will find old lock with some bad data and will SIGSEGV when reading
the owner.

Solution:

Going back the call trace for the TRUNCATE command, the following list
is an approximate order of operations:

- OPEN table (SQL code)
- LOCK table (1) (SQL code)
- CLOSE table (NDB code)
- TRUNCATE table, i.e., DROP/CREATE (2) (NDB code)
- OPEN table (3) (NDB code)
- UNLOCK table (4) (SQL code)
- CLOSE TABLE (SQL code)

The Table is already opened before TRUNCATE of SE is
called. Therefore, if CLOSE and OPEN around
TRUNCATE (ha_ndbcluster::create) can be ignored, then the LOCK
structure remains untouched. The Table is already CLOSED when
UNLOCKED, so it is expected that the resources the HANDLER has
established when OPEN (buffers, etc) are to be freed.

More specifically:

On success:
- DROP will mark the existing (locked) NDB_SHARE as Dropped (not yet physically released as the session has a reference to it)
- OPEN will create a new NDB_SHARE (2 shares will exist)
- UNLOCK will remove locks from old NDB_SHARE
- CLOSE will reduce ref count on old NDB_SHARE decrementing to
zero + freeing it (1 share exists)

On failure:
- DROP will do nothing
- UNLOCK will remove locks from old NDB_SHARE
- CLOSE will reduce ref count on old NDB_SHARE, but it will be
retained (1 share exists)

Change-Id: Ib2cbb3cf49ceac6ad2717e000b5b89beb026e4e4