Bug#36335695 error on second execution of deformed window query in function
During the first execute of a stored function, we encounter a
resolving error in a query contained in a SET variable subquery. The
LEX is not marked as broken, normally done using
Prepare_error_tracker. Therefore on the second execute, there is no
proper cleanup; we head straight to execute (exec_core) since we think
there is no issue, in reset_lex_and_exec_core (error == false)

Since items are not cleaned up, we end up in a server exit (symptom:
Item_field::fixed == true, but its ::field is nullptr).

If the LEX were marked as broken, error would be set to true, and we
would return to validate_lex_and_execute_core which does the necessary
cleanup of items for reprepare, but we never get there in the second
execution.

Solution: add Prepare_error_tracker when we decide to reprepare in
sql_set_variables.

Change-Id: Ib99c1c640cfb87d1c9fa99b22b8a081b9781fa30