Bug#37534388 PB2: ndb_rpl.rpl_vector fails in Windows
Test ndb_rpl.rpl_vector failed since it was expecting 11 lines in binlog
to mention VECTOR while only 7 was found. This since NDB by default do
not write before row image for updates and deletes to binlog as InnoDB
does. Fix is to make test aware of that before images are not written.

A bug with MTR and assert_grep.inc caused the failing part of test to
not being checked at all on platforms where environment variables have
case sensitive names.

Change-Id: I34fdf78efe32d324cdbb415e3f5b4b0ce3ef88ee