Bug#37552796 - MRS session is service specific, but it should be auth-app specific
Description
===========
MySQL REST Service session is generated after accessing an endpoint that lies under
service, for example:

    /service1/authentication/login

This may lead to an impression that, the session is service specific, but the session
is generated for an authentication_application. Each authentication application may be
assigned to multiple services. In other words if two services share a auth-app, and it
had generated a session, then the two service should be able to use that session to
grant access to the rest-user.

Current implementation creates the session for a single service.

Fix
===
Assing the session to auth-app, not to the service.

Change-Id: I0a7db67276d5c9e02e80e4f79c309c4920405566