WL#16837 Table options for Lakehouse engine attributes
Add new table options for CREATE/ALTER TABLE of external tables.
The table options will be mapped to ENGINE_ATTRIBUTE settings.

ALTER TABLE:
Distinguish between the existing ENGINE_ATTRIBUTE and if a new one is
specified in the current CREATE/ALTER statement, and only report
conflicts with the latter.

The parts of the existing ENGINE_ATTRIBUTE that are modified in the
current statement will be removed.  The existing code for CREATE TABLE
will then generate a new JSON value for this option and put into
ENGINE_ATTRIBUTE.

Special care must be taken to preserve "stand-alone" table options
that are stored inside the dialect object (e.g., ALLOW_MISSING_FILES).
Any such settings in the existing ENGINE_ATTRIBUTE will be added to
the set of table options for the current ALTER statements if not
already present.

SHOW CREATE TABLE:
The new SQL syntax will be generated from the stored ENGINE_ATTRIBUTE.
Only print ENGINE_ATTRIBUTE if the debug flag "print_engine_attribute"
is set.  The exception is JSON keys that are not supported by SQL
(i.e., region/namespace/buckets). If those are used, ENGINE_ATTRIBUTE
will then be used for specification of all files.
Mask par-id in PAR URLs.
Test that the result of SHOW CREATE TABLE will recreate an identical table.

Column specific formats:
Support EXTERNAL_FORMAT clause in column definitions for
temporal columns. Specified string will be persisted in the
ENGINE_ATTRIBUTE of the column.

Misc:
Added new error message, ER_EXTERNAL_TABLE_INVALID_OPTION,

If FILES table option is used, register the statement for rewrite so
that the PAR ids are not exposed.

The "uri" keyword in ENGINE_ATTRIBUTE may be used for PARs that are
not supported by the "par" keyword.  On the other hand, the "uri"
keyword does not support specifying both a URI and a file
name/pattern/prefix.The solution is to only use "par" when
FILE_NAME/FILE_PATTERN/ FILE_PREFIX is also specified. Otherwise,
"uri" will be used; also for PARs.

Change-Id: I82a6bd7a0ce72c30a43f6752289b28ff267dd8e4