Bug#37591520: Minor cleanups in overflow_bitset.h [1/5, noclose]
Remove the IdentityCombine and AndCombine structs, and replace them
with with std::identity and std::bit_and in OverflowBitsetBitsIn.

When constructing an OverflowBitsetBitsIn, pass the bitsets in an
array instead of in an initializer_list which is implicitly converted
to an array, so that the template parameter N can be automatically
deduced from the array size. This simplifies the calls to the
constructor a bit.

Also run include-what-you-use on overflow_bitset.{h,cc}.

Change-Id: Id04eb3b639c807a8fa1b7937ffbbfc1d43fcae34