Bug#37591520: Minor cleanups in overflow_bitset.h [2/5, noclose]
Remove OverflowBitset::empty(). One might expect that it returns true
if all bits of a bitset are unset. But it only returns true for
inlined bitsets with all bits unset. It always returns false if the
OverflowBitset is not inlined.

To avoid confusion, the function is removed, and the IsEmpty()
function is used instead.

In MakeSortPathWithoutFilesort(), empty() was used intentionally, so
an additional checks for is_inline() is added for now. These checks
will be removed by a refactoring in a later patch in this patch
series (patch 5).

Change-Id: I4a2dc13378319f109eb301036bfec1c6dea90fdf