Bug#37452558 Assertion `path->num_output_rows() <=                 path->num_output_rows_before_filter' failed.
During distinct processing in access path creation, it is expected not
to have any remaining filter conditions present in the path. An integrity
check is present for the scenarios where filter_predicates is present.
And verifies the row count integrity numbers with respect before and
after filter application. The filter_predicates variable is also used to
represent join predicates in certain scenarios. The row count check in
this scenario is not valid. Which can lead to the debug check to fail.

Fix: Update the code to limit the row count integrity check only
when the variable contains valid filter predicates.

Change-Id: I12670fd643e307c6b5301091b899c304378d746b