Bug#37510755 Performance regression in join for tables with many columns
Description:
  When determining the offsets of each field in the record, the state of
the record (w.r.t. row version) will determine which fields are present
in the record. Currently, we conditinally check the records inserted
state for every field. This causes performance delays as the number of
columns increases and this delay is significant as the offsets are
calculated for every field in every record being processed.
  Changing the order of the loops such that the base case is handled
seperately (in @rec_init_offsets_comp_ordinary) improves the
performance. The base case refers to a record which was inserted when
the table contained neither instant version nor row version. The
improvement is significant on tables with a large number of columns.

Fix:
  The case of @INSERTED_INTO_TABLE_WITH_NO_INSTANT_NO_VERSION is removed
out of the switch clause and handled in an if. The rest of the switch is
moved to the else.

Thanks to Bin Wang (wangbin579) for the contribution.

Change-Id: I92bd7d74981502e6296b0af8c1c0e9cd90fa4529