Bug#37478594  Virtual column value may be incorrectly set to NULL on cascade update
Problem:
When an ON DELETE SET NULL cascade updates a row in a child table,
setting to NULL all the columns on which a virtual column is based,
one of 2 things may happen:
1. the virtual field is also set to NULL in the update, instead of being
re-calculated.
2. The value of the virtual field is not re-calculated at all.
This impacts indexes based on the virtual field, leading to index
corruption.

Root causes:
For both ON DELETE SET NULL and ON UPDATE SET NULL actions the virtual
field calculation attempts to use parent table update vector to detrmine
which base fields should be treated as NULL for the purpose of the
calculation. However, in the ON DELETE SET NULL scenario the parent row
is deleted and consequently there is no meaningful update vector for it.
Additionally, the scenario where the virtual column is based on all the
columns that comprise the FK is treated as special, somewhat hiding the
problem - the virtual field is "calculated" to NULL without actually
calculating the underlying expression. If the expression is such that
its value is NULL if all column operands are NULL, this happens to be
correct - but obviously this not true for all expressions.

Fix:
When virtual fields are calculated as part of a FOREIGN KEY action,
all the updates for the non-virtual columns of the child table are
already known at that point. Instead of looking up updates to the
parent table, these updates may be used directly by the calculation.
This may be applied uniformly to all FK update actions: ON DELETE
SET NULL, ON UPDATE CASCADE and ON UPDATE SET NULL.

Change-Id: I6ee5c0502dc397276edac09513d0e2f27387aeaf