Bug#37591520: Minor cleanups in overflow_bitset.h [5/5]
MakeSortPathWithoutFilesort() has a check for
AccessPath::applied_sargable_join_predicates() being inline and empty.
The purpose of the check is to find out whether it is called during
join planning or after join planning. During join planning, the
applied_sargable_join_predicates bitset needs to be maintained. After
join planning, it should be ignored, as it's no longer needed and
often not initialized.

This patch removes this check by lifting the code up from
MakeSortPathWithoutFilesort() and invoke it unconditionally after the
call that is made during join planning. The other place it is
called (when applying windows), it is not needed, so the code is not
called there.

The changed also meant the num_where_predicates parameter of
MakeSortPathWithoutFilesort() was no longer in use, so it was removed,
both from this function and from some functions higher up in the call
stack.

The code that was lifted up, was for clearing the filter predicates
and retaining the applied sargable join predicates in
AccessPath::filter_predicates. This is a common operation, so the
helper function ClearFilterPredicates() was added to enable code
reuse.

One small unrelated change: While modifying the parameter list of
ApplyWindowFunctions(), clang-tidy suggested making root_candidates a
constant reference to avoid unnecessary copying or moving. The patch
makes the suggested change.

Change-Id: I08131878b9ca8ecd48cb2440302d12f2549a0779